# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive development toolkit for Odoo environments with profiling, monitoring, and optimization features.",
        "title": "OdooDevTools API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://example.com/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/environments/{env_id}/api-keys": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return all API keys for the given environment. Raw key values are never included.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api-keys"
                ],
                "summary": "List API keys",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID (UUID)",
                        "name": "env_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIKeyListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a new agent API key for the given environment. The raw key is returned once and never shown again. If expires_at is omitted the key expires in 90 days. If scopes is omitted it defaults to [\"agent:write\"].",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api-keys"
                ],
                "summary": "Create API key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID (UUID)",
                        "name": "env_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Key options",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAPIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.APIKeyCreatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/environments/{env_id}/api-keys/{key_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark an API key as inactive. The agent using it will receive 401 on the next request.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api-keys"
                ],
                "summary": "Revoke API key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID (UUID)",
                        "name": "env_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API Key ID (UUID)",
                        "name": "key_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/api.ErrorCode"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorDetail"
                    }
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "api.ErrorCode": {
            "type": "string",
            "enum": [
                "VALIDATION_ERROR",
                "UNAUTHORIZED",
                "FORBIDDEN",
                "NOT_FOUND",
                "CONFLICT",
                "UNPROCESSABLE_ENTITY",
                "RATE_LIMITED",
                "BAD_REQUEST",
                "PAYLOAD_TOO_LARGE",
                "INTERNAL_ERROR",
                "SERVICE_UNAVAILABLE",
                "TIMEOUT",
                "DATABASE_ERROR",
                "TENANT_NOT_FOUND",
                "ENVIRONMENT_NOT_FOUND",
                "AGENT_NOT_CONNECTED",
                "AGENT_NOT_FOUND",
                "QUOTA_EXCEEDED",
                "PLAN_LIMIT_REACHED",
                "INVALID_API_KEY",
                "EXPIRED_API_KEY",
                "INVALID_TOKEN",
                "EXPIRED_TOKEN",
                "INVALID_CREDENTIALS",
                "EMAIL_ALREADY_EXISTS",
                "SLUG_ALREADY_EXISTS",
                "INVALID_SCOPE",
                "SCHEMA_SNAPSHOT_NOT_FOUND",
                "MIGRATION_SCAN_NOT_FOUND",
                "ERROR_GROUP_NOT_FOUND",
                "ALERT_RULE_NOT_FOUND",
                "ALERT_CHANNEL_NOT_FOUND",
                "PROFILER_RECORDING_NOT_FOUND",
                "ANON_PROFILE_NOT_FOUND",
                "ANON_JOB_NOT_FOUND",
                "INVALID_ODOO_VERSION",
                "INVALID_DOMAIN"
            ],
            "x-enum-comments": {
                "ErrCodeExpiredAPIKey": "nolint:gosec // not a credential",
                "ErrCodeInvalidAPIKey": "nolint:gosec // not a credential",
                "ErrCodeInvalidCredentials": "nolint:gosec // not a credential"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "nolint:gosec // not a credential",
                "nolint:gosec // not a credential",
                "",
                "",
                "nolint:gosec // not a credential",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                ""
            ],
            "x-enum-varnames": [
                "ErrCodeValidation",
                "ErrCodeUnauthorized",
                "ErrCodeForbidden",
                "ErrCodeNotFound",
                "ErrCodeConflict",
                "ErrCodeUnprocessable",
                "ErrCodeRateLimited",
                "ErrCodeBadRequest",
                "ErrCodePayloadTooLarge",
                "ErrCodeInternal",
                "ErrCodeUnavailable",
                "ErrCodeTimeout",
                "ErrCodeDatabase",
                "ErrCodeTenantNotFound",
                "ErrCodeEnvironmentNotFound",
                "ErrCodeAgentNotConnected",
                "ErrCodeAgentNotFound",
                "ErrCodeQuotaExceeded",
                "ErrCodePlanLimitReached",
                "ErrCodeInvalidAPIKey",
                "ErrCodeExpiredAPIKey",
                "ErrCodeInvalidToken",
                "ErrCodeExpiredToken",
                "ErrCodeInvalidCredentials",
                "ErrCodeEmailAlreadyExists",
                "ErrCodeSlugAlreadyExists",
                "ErrCodeInvalidScope",
                "ErrCodeSchemaSnapshotNotFound",
                "ErrCodeMigrationScanNotFound",
                "ErrCodeErrorGroupNotFound",
                "ErrCodeAlertRuleNotFound",
                "ErrCodeAlertChannelNotFound",
                "ErrCodeProfilerRecordingNotFound",
                "ErrCodeAnonProfileNotFound",
                "ErrCodeAnonJobNotFound",
                "ErrCodeInvalidOdooVersion",
                "ErrCodeInvalidDomain"
            ]
        },
        "api.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "dto.APIKeyCreatedResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "env_id": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key_prefix": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "raw_key": {
                    "description": "shown once; not stored in plaintext",
                    "type": "string"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "dto.APIKeyItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "key_prefix": {
                    "type": "string"
                },
                "last_used": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.APIKeyListResponse": {
            "type": "object",
            "properties": {
                "keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.APIKeyItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CreateAPIKeyRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255
                },
                "expires_at": {
                    "description": "RFC3339; defaults to 90 days from now if omitted",
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"ApiKey\" followed by a space and your API key (e.g. \"ApiKey odt_xxx...\").",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}